refactor: v0.4.0-alpha.5 (#128)
* remove old configs, use single config

* bump versions

* use local `april-21-unified-config` docker image

* note

* rename config -> wavs

* bump versions

* refactor: default networks

* fix: all http requests

* working entire flow

* use ghcr.io/reecepbcups/wavs:april-23-0.4.0-alpha.5

* use the -1 approach upstream

* wavs middleware official image + latest git version

* chore(deps): 0.4.0-alpha.5

* reduced README complexity

* fix: go runner

* fix: js

* comment out docci for now

* middleware proper tag

+++ script/start_all.sh
@@ -3,13 +3,14 @@
 set -e
 
 PORT=8545
-MIDDLEWARE_IMAGE=ghcr.io/reecepbcups/wavs-middleware:0.0.1
+MIDDLEWARE_IMAGE=ghcr.io/lay3rlabs/wavs-middleware:0.4.0-alpha.5
 LOG_FILE=.docker/start.log
+OPERATOR_PK=${OPERATOR_PK:-""}
+OPERATOR_MNEMONIC=${OPERATOR_MNEMONIC:-""}
 export DOCKER_DEFAULT_PLATFORM=linux/amd64
 
 ## == Start watcher ==
 rm $LOG_FILE 2> /dev/null || true
-
 
 ## == Base Anvil Testnet Fork ==
 anvil --fork-url https://ethereum-holesky-rpc.publicnode.com --port ${PORT} &
@@ -20,27 +21,32 @@ do
   sleep 0.25
 done
 
+if [[ -z "$OPERATOR_PK" ]]; then
+  echo "You must set OPERATOR_PK"
+  exit 1
+fi
+if [[ -z "$OPERATOR_MNEMONIC" ]]; then
+  echo "You must set OPERATOR_MNEMONIC"
+  exit 1
+fi
 
 ## == Eigenlayer ==
 # setup
 docker run --rm --network host --env-file .env -v ./.nodes:/root/.nodes "$MIDDLEWARE_IMAGE"
-PRIVATE_KEY=$(cat .nodes/deployer)
-# operator register
-docker run --rm --network host --env-file .env -v ./.nodes:/root/.nodes --entrypoint /wavs/register.sh "$MIDDLEWARE_IMAGE" "$PRIVATE_KEY"
 
 
 ## == Setup Deployer
-echo "Using Address: $(cast wallet address --private-key ${PRIVATE_KEY})"
+echo "Using Address: $(cast wallet address --private-key ${OPERATOR_PK})"
 
 SP=""
 if [[ "$(uname)" == *"Darwin"* ]]; then
   SP=" "
 fi
 
-sed -i${SP}'' -e "s/^WAVS_CLI_ETH_MNEMONIC=.*$/WAVS_CLI_ETH_MNEMONIC=\"$PRIVATE_KEY\"/" .env
-sed -i${SP}'' -e "s/^WAVS_SUBMISSION_MNEMONIC=.*$/WAVS_SUBMISSION_MNEMONIC=\"$PRIVATE_KEY\"/" .env
-sed -i${SP}'' -e "s/^WAVS_AGGREGATOR_MNEMONIC=.*$/WAVS_AGGREGATOR_MNEMONIC=\"$PRIVATE_KEY\"/" .env
-
+sed -i${SP}'' -e "s/^WAVS_CLI_ETH_CREDENTIAL=.*$/WAVS_CLI_ETH_CREDENTIAL=\"$OPERATOR_PK\"/" .env
+sed -i${SP}'' -e "s/^WAVS_AGGREGATOR_CREDENTIAL=.*$/WAVS_AGGREGATOR_CREDENTIAL=\"$OPERATOR_PK\"/" .env
+# this is what we generate other addresses in service manager based off of.
+sed -i${SP}'' -e "s/^WAVS_SUBMISSION_MNEMONIC=.*$/WAVS_SUBMISSION_MNEMONIC=\"$OPERATOR_MNEMONIC\"/" .env
 
 # == WAVS & Aggregator ==
 docker compose up --remove-orphans &
